interface for sv3 (serial version 3)
The is an interface for sv3 (serial version 3) devices of which the POnn series and BV460n motor controllers use. It is a set of functions common to all of these devices and mostly involve setting and reading the EEPROM information within these devices. A useful function is when there is more than one device on the serial but,it can discover which devices are connected by their address.

sv3a.py can be configured for I2C or Serial, it establishes a 'bus'. One the bus has been established it works the same for either serial or I2C even though the atual electrical connections may be different.

In a nutshell, the sv3a.py provides the link between the different devices and the hardware whether it be Serial or I2C.